+++ app_meta.js
@@ -1,4 +1,4 @@
 const AppMeta = {
-  version: "0.4.8.7",
-  date: "Dec 15 2025"
+  version: "0.4.8.8",
+  date: "Dec 17 2025"
 };